+++ codey/knowledge/__init__.py
@@ -2,15 +2,18 @@
 
 from codey.knowledge.brief import KnowledgeBriefBuilder, ResearchBrief
 from codey.knowledge.changes import KnowledgeChanges, RestoreResult
+from codey.knowledge.graph import KnowledgeGraphBuilder, ResearchGraphArtifact
 from codey.knowledge.note import NOTE_TYPES, KnowledgeNote
 from codey.knowledge.store import KnowledgeStore
 
 __all__ = [
     "KnowledgeBriefBuilder",
     "KnowledgeChanges",
+    "KnowledgeGraphBuilder",
     "KnowledgeNote",
     "KnowledgeStore",
     "NOTE_TYPES",
     "ResearchBrief",
+    "ResearchGraphArtifact",
     "RestoreResult",
 ]